requirements.txt
asgiref==3.7.2
Django==4.1.11
django-taggit==3.0.0
Markdown==3.4.4
psycopg2-binary==2.9.3
sqlparse==0.4.4
typing_extensions==4.7.1
